Refactor useStickyPosition to init when scrolled away from top of viewport
